// src/modules/profit_and_loss/translate.enum.ts
export enum ActualBudgetItemKeys {
    NOT_FOUND = "actualBudgetItem.NOT_FOUND",
    CREATION_ERROR = "actualBudgetItem.CREATION_ERROR",
    GET_Months_ERROR = "actualBudgetItem.GET_Months_ERROR",
    UPDATED_SUCCESSFULLY = "actualBudgetItem.UPDATED_SUCCESSFULLY",
    COMPANY_ACCESS_DENIED = "actualBudgetItem.COMPANY_ACCESS_DENIED",
    DELETED_SUCCESSFULLY = "actualBudgetItem.DELETED_SUCCESSFULLY",
}

export enum ActualBudgetMonthKeys {
    NOT_FOUND = "ActualBudgetMonth.NOT_FOUND",
}