// src/modules/company/translate.enum.ts
export enum CompanyKeys {
    NOT_FOUND = "company.NOT_FOUND",
    CREATION_ERROR = "company.CREATION_ERROR",
    UPDATE_ERROR = "company.UPDATE_ERROR",
    UPDATED_SUCCESSFULLY = "company.UPDATED_SUCCESSFULLY",
    CANT_DELETE = "company.CANT_DELETE",
    DELETED_SUCCESSFULLY = "company.DELETED_SUCCESSFULLY",
    PLAN_DATES_NOT_FOUND = "company.PLAN_DATES_NOT_FOUND",
    FINANCIAL_YEARS_ERROR = "company.FINANCIAL_YEARS_ERROR",
    PLAN_DATES_CREATION_ERROR = "company.PLAN_DATES_CREATION_ERROR",
    PLAN_DATES_UPDATE_ERROR = "company.PLAN_DATES_UPDATE_ERROR",
    PLAN_DATES_UPDATED_SUCCESSFULLY = "company.PLAN_DATES_UPDATED_SUCCESSFULLY",
    INVALID_PLAN_DATES = "company.INVALID_PLAN_DATES",
    ORG_CHART_ERROR = "company.ORG_CHART_ERROR",
}

export enum InvitationKeys {
    NOT_FOUND = "invitation.NOT_FOUND",
    REGISTERED_STARTUP_EMAIL_ERROR = "invitation.REGISTERED_STARTUP_EMAIL_ERROR",
    SEND_ERROR = "invitation.SEND_ERROR",
    SENT_SUCCESSFULLY = "invitation.SENT_SUCCESSFULLY",
    UPDATE_ERROR = "invitation.UPDATE_ERROR",
    UPDATED_SUCCESSFULLY = "invitation.UPDATED_SUCCESSFULLY",
    CANT_DELETE = "invitation.CANT_DELETE",
    DELETED_SUCCESSFULLY = "invitation.DELETED_SUCCESSFULLY",
}

export enum InvestorRelationKeys {
    NOT_FOUND = "investorRelation.NOT_FOUND",
    UPDATE_ERROR = "investorRelation.UPDATE_ERROR",
    UPDATED_SUCCESSFULLY = "investorRelation.UPDATED_SUCCESSFULLY",
    CANT_DELETE = "investorRelation.CANT_DELETE",
    DELETED_SUCCESSFULLY = "investorRelation.DELETED_SUCCESSFULLY",
}